test(e2e): hoist openAddAccountModal to shared-flows + harden Redux assertion (#773)
- Move `openAddAccountModal` into `app/test/e2e/helpers/shared-flows.ts`;
  Slack and WhatsApp specs now import the single shared definition instead
  of carrying byte-identical copies of the helper.
- Replace the dead `registered === undefined` fallback in both specs with
  `browser.waitUntil` over the Redux selector. The previous check was
  unreachable: `browser.execute` returns the callback's value (always
  boolean), so `=== undefined` never matched and the fallback never fired.
  With the store exposed on `window.__OPENHUMAN_STORE__`, the wait now
  fails loudly if registration never lands instead of silently passing.
