+++ app/test/e2e/specs/slack-flow.spec.ts
@@ -8,28 +8,12 @@ import {
   waitForWindowVisible,
 } from '../helpers/element-helpers';
 import { supportsExecuteScript } from '../helpers/platform';
-import { completeOnboardingIfVisible, navigateViaHash } from '../helpers/shared-flows';
+import {
+  completeOnboardingIfVisible,
+  navigateViaHash,
+  openAddAccountModal,
+} from '../helpers/shared-flows';
 import { startMockServer, stopMockServer } from '../mock-server';
-
-async function openAddAccountModal(): Promise<void> {
-  // The "Add app" affordance only carries an `aria-label` — its visible text
-  // sits inside a `pointer-events: none` tooltip span, so the shared
-  // `clickButton`/`clickText` helpers can't target it. Tracking a
-  // `clickByAriaLabel` helper as a follow-up.
-  const opened = await browser.execute(() => {
-    const buttons = Array.from(document.querySelectorAll<HTMLButtonElement>('button'));
-    const addBtn = buttons.find(b => b.getAttribute('aria-label') === 'Add app');
-    if (addBtn) {
-      addBtn.click();
-      return true;
-    }
-    return false;
-  });
-  if (!opened) {
-    throw new Error('Could not locate Add app button on /accounts');
-  }
-  await waitForText('Add account', 5_000);
-}
 
 /**
  * Smoke spec for the Slack account integration (feature 10.1.4).
@@ -111,19 +95,25 @@ describe('Slack account integration smoke', () => {
     // 2) Redux must record a new account with provider === "slack" — the
     // backing-state mock-effect that proves registration. The Slack tile
     // label and the post-pick rail tooltip share the literal string "Slack",
-    // so a pure DOM assertion cannot distinguish them.
-    const registered = await browser.execute(() => {
-      const winAny = window as unknown as { __OPENHUMAN_STORE__?: { getState: () => unknown } };
-      const state = winAny.__OPENHUMAN_STORE__?.getState() as
-        | { accounts?: { accounts?: Record<string, { provider?: string }> } }
-        | undefined;
-      const accounts = state?.accounts?.accounts ?? {};
-      return Object.values(accounts).some(a => a.provider === 'slack');
-    });
-    if (registered === undefined) {
-      expect(await textExists('Slack')).toBe(true);
-    } else {
-      expect(registered).toBe(true);
-    }
+    // so a pure DOM assertion cannot distinguish them. The store handle is
+    // exposed on `window.__OPENHUMAN_STORE__` from `app/src/store/index.ts`.
+    const registered = await browser.waitUntil(
+      async () =>
+        await browser.execute(() => {
+          const winAny = window as unknown as { __OPENHUMAN_STORE__?: { getState: () => unknown } };
+          const state = winAny.__OPENHUMAN_STORE__?.getState() as
+            | { accounts?: { accounts?: Record<string, { provider?: string }> } }
+            | undefined;
+          if (!state) return false;
+          const accounts = state.accounts?.accounts ?? {};
+          return Object.values(accounts).some(a => a.provider === 'slack');
+        }),
+      {
+        timeout: 5_000,
+        timeoutMsg:
+          'Redux accounts slice never recorded a slack provider after picking the Slack tile',
+      }
+    );
+    expect(registered).toBe(true);
   });
 });
